+++ Magic/CHANGELOG.md
@@ -10,6 +10,7 @@
  - Fix shops with inline item name/lore definitions
  - Some fixes to the ModifyProperties action
  - Add CheckHealth action
+ - Allow actions on SpawnEntity which will run when the entity dies (See: kamikaze spell)
  - Allow multiple commands in a command NPC via the ; delimiter
 
 # 7.6